requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
customtkinter==5.2.1
darkdetect==0.8.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.11
packaging==24.0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.0.7
requests==2.31.0
sqlparse==0.5.5
urllib3==2.6.3
PyPDF2==3.0.1
